pyproject.toml
## Project requirements (PEP 518, https://python-poetry.org/).


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]


[tool.black]
line-length = 96
target-version = ["py39", "py310", "py311", "py312"]


[tool.isort]
line_length = 96
profile = "black"


[tool.mypy]
explicit_package_bases = true
namespace_packages = true

check_untyped_defs = true
strict_optional = true
no_implicit_optional = true

warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true

color_output = false
pretty = true
show_error_codes = true

incremental = true
sqlite_cache = true


[tool.poetry]
name = "s3_tools"
version = "1.1.0"
description = "Assorted utilities for working with S3 storage"
license = "Apache-2.0"
authors = [
    "Brian Aydemir <baydemir@morgridge.org>",
]
repository = "https://github.com/brianaydemir/s3-tools"
packages = [{ include = "app" }]


[tool.poetry.dependencies]
python = "^3.9"
humanize = "~4.10"
minio = "~7.2"
radosgw-admin = "~1.7"


[tool.poetry.group.dev.dependencies]
bandit = "~1.7"
black = "~24.8"
isort = "~5.13"
markdown = "~3.7"
mypy = "~1.11"
pre-commit = "~3.8"
pylint = "~3.3"


[tool.pylint.messages-control]
max-line-length = 96
disable = "invalid-name,missing-class-docstring,missing-function-docstring"
